LinkedIn/Linkedin_posts/requirements.txt
selenium
webdriver-manager
sentence-transformers
spacy
nltk
torch
transformers
huggingface-hub
pandas
praw
python-dotenv
PyGithub
python-dotenv
langdetect
BeautifulSoup4

# python -m spacy download en_core_web_sm